Fixed Crash, Moved Debug Pos, And Patched BitE Crash
Fixed the crash that was happening because of the background music; a reimplementation later is needed. Moved the position for the debug position, angle and speed to a better place for wii. Changed the angle print to hexadecimal instead of decimal. Implemented a patch for the vanilla BitE crash (which just checks if your on the title screen map).